Orchestration flags as separate class, ServiceRequestForm builder now validates after building, and ServiceRequestForm::set* methods removed to force users to build using the builder and thus use validated forms.
