+++ common/src/main/java/eu/arrowhead/common/model/ServiceRequestForm.java
@@ -18,46 +18,6 @@ import java.util.*;
  */
 public class ServiceRequestForm {
 
-  public enum Flags {
-    TRIGGER_INTER_CLOUD("triggerInterCloud"),
-    EXTERNAL_SERVICE_REQUEST("externalServiceRequest"),
-    ENABLE_INTER_CLOUD("enableInterCloud"),
-    METADATA_SEARCH("metadataSearch"),
-    PING_PROVIDERS("pingProviders"),
-    OVERRIDE_STORE("overrideStore"),
-    MATCHMAKING("matchmaking"),
-    ONLY_PREFERRED("onlyPreferred"),
-    ENABLE_QOS("enableQoS"),
-    ;
-
-    private final String flag;
-
-    Flags(final String flag) {
-      this.flag = flag;
-    }
-
-    @Override
-    public String toString() {
-      return flag;
-    }
-  }
-
-  public static class OrchestrationFlags extends HashMap<String, Boolean> {
-    public OrchestrationFlags() {
-      for (Flags flag : Flags.values()) {
-        put(flag, false);
-      }
-    }
-
-    public Boolean get(Flags flag) {
-      return get(flag.toString());
-    }
-
-    public Boolean put(Flags flag, Boolean value) {
-      return super.put(flag.toString(), value);
-    }
-  }
-
   private ArrowheadSystem requesterSystem;
   private ArrowheadCloud requesterCloud;
   private ArrowheadService requestedService;
@@ -83,24 +43,12 @@ public class ServiceRequestForm {
     return requesterSystem;
   }
 
-  public void setRequesterSystem(ArrowheadSystem requesterSystem) {
-    this.requesterSystem = requesterSystem;
-  }
-
   public ArrowheadCloud getRequesterCloud() {
     return requesterCloud;
   }
 
-  public void setRequesterCloud(ArrowheadCloud requesterCloud) {
-    this.requesterCloud = requesterCloud;
-  }
-
   public ArrowheadService getRequestedService() {
     return requestedService;
-  }
-
-  public void setRequestedService(ArrowheadService requestedService) {
-    this.requestedService = requestedService;
   }
 
 
@@ -108,44 +56,16 @@ public class ServiceRequestForm {
     return orchestrationFlags;
   }
 
-  public void setOrchestrationFlags(OrchestrationFlags orchestrationFlags) {
-    orchestrationFlags.forEach((k, v)->this.orchestrationFlags.put(k, v));
-  }
-
-  public void setOrchestrationFlags(Map<String, Boolean> orchestrationFlags) {
-    orchestrationFlags.forEach((k, v)->this.orchestrationFlags.put(k, v));
-  }
-
-  public void setFlag(Flags flag, Boolean value) {
-    this.orchestrationFlags.put(flag, value);
-  }
-
-  public void setFlag(String key, Boolean value) {
-    this.orchestrationFlags.put(key, value);
-  }
-
   public List<PreferredProvider> getPreferredProviders() {
     return preferredProviders;
-  }
-
-  public void setPreferredProviders(List<PreferredProvider> preferredProviders) {
-    this.preferredProviders = preferredProviders;
   }
 
   public Map<String, String> getRequestedQoS() {
     return requestedQoS;
   }
 
-  public void setRequestedQoS(Map<String, String> requestedQoS) {
-    this.requestedQoS = requestedQoS;
-  }
-
   public Map<String, String> getCommands() {
     return commands;
-  }
-
-  public void setCommands(Map<String, String> commands) {
-    this.commands = commands;
   }
 
   public static class Builder {
@@ -175,13 +95,23 @@ public class ServiceRequestForm {
       return this;
     }
 
-    public Builder flag(Flags flag, Boolean value) {
+    public Builder flag(OrchestrationFlags.Flags flag, Boolean value) {
       this.orchestrationFlags.put(flag, value);
       return this;
     }
 
     public Builder flag(String key, Boolean value) {
       this.orchestrationFlags.put(key, value);
+      return this;
+    }
+
+    public Builder flags(Map<String, Boolean> flags) {
+      this.orchestrationFlags.putAll(flags);
+      return this;
+    }
+
+    public Builder flags(OrchestrationFlags flags) {
+      this.orchestrationFlags.putAll(flags);
       return this;
     }
 
@@ -201,16 +131,16 @@ public class ServiceRequestForm {
     }
 
     public ServiceRequestForm build() {
-      return new ServiceRequestForm(this);
+      return new ServiceRequestForm(this).validateCrossParameterConstraints();
     }
   }
 
-  public void validateCrossParameterConstraints() {
-    if (requestedService == null && orchestrationFlags.get(Flags.OVERRIDE_STORE)) {
+  public ServiceRequestForm validateCrossParameterConstraints() {
+    if (requestedService == null && orchestrationFlags.get(OrchestrationFlags.Flags.OVERRIDE_STORE)) {
       throw new BadPayloadException("RequestedService can not be null when overrideStore is TRUE");
     }
 
-    if (orchestrationFlags.get(Flags.ONLY_PREFERRED)) {
+    if (orchestrationFlags.get(OrchestrationFlags.Flags.ONLY_PREFERRED)) {
       List<PreferredProvider> tmp = new ArrayList<>();
       for (PreferredProvider provider : preferredProviders) {
         if (!provider.isValid()) {
@@ -223,9 +153,11 @@ public class ServiceRequestForm {
       }
     }
 
-    if (orchestrationFlags.get(Flags.ENABLE_QOS) && (requestedQoS.isEmpty() || commands.isEmpty())) {
+    if (orchestrationFlags.get(OrchestrationFlags.Flags.ENABLE_QOS) && (requestedQoS.isEmpty() || commands.isEmpty())) {
       throw new BadPayloadException("RequestedQoS or commands hashmap is empty while \"enableQoS\" is set to true");
     }
+
+    return this;
   }
 
 }
